Add dynamic usage mode selection for GPS Position.
Most of the GPS chips we support have the ability to change their
operating model "dynamics" based on the intended activities.

We currently assume everyone is a pedestrian and configure as such.

However, there are users out there on bikes, planes, balloons, and
with fixed devices. Where GPS chips support it, we can configure for
each of this circumstances and improve performance.

For example, fixed nodes might no longer jump around on the map.
Planes will have a more responsive and less filtered position.
Balloons' GPS won't simply cut out because its gone over an altitude
limit. Boats won't have altitudes that make them appear to be floating
in the air.

Additionally, this could provide the opportunity for a bit of fun
in the apps - surface information about the use case so people know
when they're connecting to a plane or baloon :)

Related to: https://github.com/meshtastic/firmware/issues/3984
Related to: https://github.com/meshtastic/protobufs/pull/619
Related to: https://github.com/meshtastic/firmware/pull/5281